+++ packages/datastore/__tests__/helpers/index.ts
@@ -1,0 +1,5 @@
+import { __modelMeta__ } from '../../src/types';
+import { initSchema as _initSchema } from '../../src/datastore/datastore';
+export * from './schemas';
+export { getDataStore } from './datastoreFactory';
+export * from './util';
