+++ packages/datastore/__tests__/helpers/schemas/internal.ts
@@ -1,0 +1,332 @@
+import { InternalSchema } from '../../../src/types';
+
+export function internalTestSchema(): InternalSchema {
+	return {
+		namespaces: {
+			datastore: {
+				name: 'datastore',
+				relationships: {
+					Setting: {
+						indexes: [],
+						relationTypes: [],
+					},
+				},
+				enums: {},
+				nonModels: {},
+				models: {
+					Setting: {
+						name: 'Setting',
+						pluralName: 'Settings',
+						syncable: false,
+						fields: {
+							id: {
+								name: 'id',
+								type: 'ID',
+								isRequired: true,
+								isArray: false,
+							},
+							key: {
+								name: 'key',
+								type: 'String',
+								isRequired: true,
+								isArray: false,
+							},
+							value: {
+								name: 'value',
+								type: 'String',
+								isRequired: true,
+								isArray: false,
+							},
+						},
+					},
+				},
+			},
+			user: {
+				name: 'user',
+				enums: {},
+				models: {
+					Model: {
+						name: 'Model',
+						pluralName: 'Models',
+						syncable: true,
+						fields: {
+							id: {
+								name: 'id',
+								isArray: false,
+								type: 'ID',
+								isRequired: true,
+							},
+							field1: {
+								name: 'field1',
+								isArray: false,
+								type: 'String',
+								isRequired: true,
+							},
+							optionalField1: {
+								name: 'optionalField1',
+								isArray: false,
+								type: 'String',
+								isRequired: false,
+							},
+							dateCreated: {
+								name: 'dateCreated',
+								isArray: false,
+								type: 'AWSDateTime',
+								isRequired: true,
+								attributes: [],
+							},
+							emails: {
+								name: 'emails',
+								isArray: true,
+								type: 'AWSEmail',
+								isRequired: true,
+								attributes: [],
+								isArrayNullable: true,
+							},
+							ips: {
+								name: 'ips',
+								isArray: true,
+								type: 'AWSIPAddress',
+								isRequired: false,
+								attributes: [],
+								isArrayNullable: true,
+							},
+							metadata: {
+								name: 'metadata',
+								isArray: false,
+								type: {
+									nonModel: 'Metadata',
+								},
+								isRequired: false,
+								attributes: [],
+							},
+							logins: {
+								name: 'logins',
+								isArray: true,
+								type: {
+									nonModel: 'Login',
+								},
+								isRequired: false,
+								attributes: [],
+							},
+						},
+					},
+					LocalModel: {
+						name: 'LocalModel',
+						pluralName: 'LocalModels',
+						syncable: false,
+						fields: {
+							id: {
+								name: 'id',
+								isArray: false,
+								type: 'ID',
+								isRequired: true,
+							},
+							field1: {
+								name: 'field1',
+								isArray: false,
+								type: 'String',
+								isRequired: true,
+							},
+						},
+					},
+				},
+				nonModels: {
+					Metadata: {
+						name: 'Metadata',
+						fields: {
+							author: {
+								name: 'author',
+								isArray: false,
+								type: 'String',
+								isRequired: true,
+								attributes: [],
+							},
+							tags: {
+								name: 'tags',
+								isArray: true,
+								type: 'String',
+								isRequired: false,
+								isArrayNullable: true,
+								attributes: [],
+							},
+							rewards: {
+								name: 'rewards',
+								isArray: true,
+								type: 'String',
+								isRequired: true,
+								attributes: [],
+							},
+							penNames: {
+								name: 'penNames',
+								isArray: true,
+								type: 'String',
+								isRequired: true,
+								isArrayNullable: true,
+								attributes: [],
+							},
+							nominations: {
+								name: 'nominations',
+								isArray: true,
+								type: 'String',
+								isRequired: false,
+								attributes: [],
+							},
+							misc: {
+								name: 'misc',
+								isArray: true,
+								type: 'String',
+								isRequired: false,
+								isArrayNullable: true,
+								attributes: [],
+							},
+							login: {
+								name: 'login',
+								isArray: false,
+								type: {
+									nonModel: 'Login',
+								},
+								isRequired: false,
+								attributes: [],
+							},
+						},
+					},
+					Login: {
+						name: 'Login',
+						fields: {
+							username: {
+								name: 'username',
+								isArray: false,
+								type: 'String',
+								isRequired: true,
+								attributes: [],
+							},
+						},
+					},
+				},
+				relationships: {
+					Model: {
+						indexes: [],
+						relationTypes: [],
+					},
+					LocalModel: {
+						indexes: [],
+						relationTypes: [],
+					},
+				},
+			},
+			sync: {
+				name: 'sync',
+				relationships: {
+					MutationEvent: {
+						indexes: [],
+						relationTypes: [],
+					},
+					ModelMetadata: {
+						indexes: [],
+						relationTypes: [],
+					},
+				},
+				enums: {
+					OperationType: {
+						name: 'OperationType',
+						values: ['CREATE', 'UPDATE', 'DELETE'],
+					},
+				},
+				nonModels: {},
+				models: {
+					MutationEvent: {
+						name: 'MutationEvent',
+						pluralName: 'MutationEvents',
+						syncable: false,
+						fields: {
+							id: {
+								name: 'id',
+								type: 'ID',
+								isRequired: true,
+								isArray: false,
+							},
+							model: {
+								name: 'model',
+								type: 'String',
+								isRequired: true,
+								isArray: false,
+							},
+							data: {
+								name: 'data',
+								type: 'String',
+								isRequired: true,
+								isArray: false,
+							},
+							modelId: {
+								name: 'modelId',
+								type: 'String',
+								isRequired: true,
+								isArray: false,
+							},
+							operation: {
+								name: 'operation',
+								type: {
+									enum: 'Operationtype',
+								},
+								isArray: false,
+								isRequired: true,
+							},
+							condition: {
+								name: 'condition',
+								type: 'String',
+								isArray: false,
+								isRequired: true,
+							},
+						},
+					},
+					ModelMetadata: {
+						name: 'ModelMetadata',
+						pluralName: 'ModelsMetadata',
+						syncable: false,
+						fields: {
+							id: {
+								name: 'id',
+								type: 'ID',
+								isRequired: true,
+								isArray: false,
+							},
+							namespace: {
+								name: 'namespace',
+								type: 'String',
+								isRequired: true,
+								isArray: false,
+							},
+							model: {
+								name: 'model',
+								type: 'String',
+								isRequired: true,
+								isArray: false,
+							},
+							lastSync: {
+								name: 'lastSync',
+								type: 'Int',
+								isRequired: false,
+								isArray: false,
+							},
+							lastFullSync: {
+								name: 'lastFullSync',
+								type: 'Int',
+								isRequired: false,
+								isArray: false,
+							},
+							fullSyncInterval: {
+								name: 'fullSyncInterval',
+								type: 'Int',
+								isRequired: true,
+								isArray: false,
+							},
+						},
+					},
+				},
+			},
+		},
+		version: '1',
+		codegenVersion: '3.2.0',
+	};
+}
