+++ packages/datastore/package.json
@@ -102,7 +102,7 @@
 		"testPathIgnorePatterns": [
 			"__tests__/model.ts",
 			"__tests__/schema.ts",
-			"__tests__/helpers.ts",
+			"__tests__/helpers/",
 			"__tests__/commonAdapterTests.ts"
 		],
 		"moduleFileExtensions": [
